+++ changelog.md
@@ -75,6 +75,7 @@ Codion Change Log
 - DenormalizedValueProvider now validates the denormalized attribute in case the source is a foreign key.
 - Entity.groupByValue() can now handle null values.
 - AbstractCondition now throws exception if the number of condition columns does not match the number of values.
+- Condition.custom() moved to ConditionType.get().
 ### is.codion.framework.db.core
 - EntityConnectionProvider.clientType() and clientVersion() now return Optional.
 ### is.codion.framework.db.local